Plant_Disease_detection_APP/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1

torch==2.1.2
torchvision==0.16.2
timm==0.9.12

numpy==1.26.4
pillow==10.2.0
opencv-python-headless==4.9.0.80

python-multipart

